build: upgrade core version to 0.15.3

+++ src/main/java/run/ikaros/plugin/bgmtv/listener/EpisodeCollectionFinishChangeListener.java
@@ -81,7 +81,7 @@ public class EpisodeCollectionFinishChangeListener {
     }
 
 
-    private Mono<Integer> getSubjectEpsSeq(Long episodeId, Long subjectId) {
+    private Mono<Float> getSubjectEpsSeq(Long episodeId, Long subjectId) {
         return subjectOperate.findById(subjectId)
                 .flatMapMany(subject -> Flux.fromStream(subject.getEpisodes().stream()))
                 .filter(episode -> EpisodeGroup.MAIN.equals(episode.getGroup()))
